+++ packages/cli/src/commands/deploy.ts
@@ -10,8 +10,8 @@ import { DEFAULT_IPFS_URL } from '../command-helpers/ipfs';
 import { createJsonRpcClient } from '../command-helpers/jsonrpc';
 import { updateSubgraphNetwork } from '../command-helpers/network';
 import { chooseNodeUrl, getHostedServiceSubgraphId } from '../command-helpers/node';
-import { validateStudioNetwork } from '../command-helpers/studio';
 import { assertGraphTsVersion, assertManifestApiVersion } from '../command-helpers/version';
+import { GRAPH_CLI_SHARED_HEADERS } from '../constants';
 import Protocol from '../protocols';
 
 const headersFlag = Flags.custom<Record<string, string>>({
@@ -170,7 +170,10 @@ export default class DeployCommand extends Command {
       }
 
       // @ts-expect-error options property seems to exist
-      client.options.headers = { Authorization: 'Bearer ' + deployKey };
+      client.options.headers = {
+        ...GRAPH_CLI_SHARED_HEADERS,
+        Authorization: 'Bearer ' + deployKey,
+      };
 
       const subgraphName = await ux.prompt('What is the name of the subgraph you want to deploy?', {
         required: true,
@@ -299,7 +302,10 @@ export default class DeployCommand extends Command {
     deployKey = await identifyDeployKey(node, deployKey);
     if (deployKey !== undefined && deployKey !== null) {
       // @ts-expect-error options property seems to exist
-      client.options.headers = { Authorization: 'Bearer ' + deployKey };
+      client.options.headers = {
+        ...GRAPH_CLI_SHARED_HEADERS,
+        Authorization: 'Bearer ' + deployKey,
+      };
     }
 
     // Ask for label if not on hosted service
@@ -384,7 +390,10 @@ export default class DeployCommand extends Command {
       // Connect to the IPFS node (if a node address was provided)
       const ipfsClient = create({
         url: appendApiVersionForGraph(ipfs.toString()),
-        headers,
+        headers: {
+          ...headers,
+          ...GRAPH_CLI_SHARED_HEADERS,
+        },
       });
 
       // Fetch the manifest from IPFS
@@ -400,28 +409,8 @@ export default class DeployCommand extends Command {
 
       await ipfsClient.pin.add(ipfsHash);
 
-      try {
-        const dataSourcesAndTemplates = DataSourcesExtractor.fromManifestString(manifestFile);
-
-        for (const { network } of dataSourcesAndTemplates) {
-          validateStudioNetwork({ studio, product, network });
-        }
-      } catch (e) {
-        this.error(e, { exit: 1 });
-      }
-
       await deploySubgraph(ipfsHash);
       return;
-    }
-
-    try {
-      const dataSourcesAndTemplates = await DataSourcesExtractor.fromFilePath(manifest);
-
-      for (const { network } of dataSourcesAndTemplates) {
-        validateStudioNetwork({ studio, product, network });
-      }
-    } catch (e) {
-      this.error(e, { exit: 1 });
     }
 
     let protocol;