+++ packages/cli/src/protocols/index.ts
@@ -110,6 +110,7 @@ export default class Protocol {
         'mbase',
         'arbitrum-one',
         'arbitrum-goerli',
+        'arbitrum-sepolia',
         'optimism',
         'optimism-goerli',
         'aurora',